@@ -1,21 +1,23 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <DockerTargetOS>Linux</DockerTargetOS>
     <UserSecretsId>c18a4789-8824-4ff5-9ba2-a77f4d304a6d</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="CoreHtmlToImage" Version="1.0.5" />
-    <PackageReference Include="jsreport.AspNetCore" Version="2.0.0" />
-    <PackageReference Include="jsreport.Binary" Version="2.5.0" />
-    <PackageReference Include="jsreport.Binary.Linux" Version="2.5.0" />
-    <PackageReference Include="jsreport.Local" Version="2.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.0.1916590" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.9" />
-    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="2.1.4" />
+    <PackageReference Include="jsreport.AspNetCore" Version="2.0.2" />
+    <PackageReference Include="jsreport.Binary" Version="2.6.1" />
+    <PackageReference Include="jsreport.Binary.Linux" Version="2.6.1" />
+    <PackageReference Include="jsreport.Local" Version="2.2.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="3.1.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.1" />
+    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.9.10" />
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.1" />
+    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="3.1.1" />
   </ItemGroup>
 
   <ItemGroup>
