@@ -1,29 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Runtime.InteropServices;
+﻿using EquilibriumCore.Areas.Identity.Data;
 using EquilibriumCore.Data;
 using EquilibriumCore.Models;
+using EquilibriumCore.Service;
 using jsreport.AspNetCore;
-using jsreport.Binary;
 using jsreport.Local;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Pomelo.EntityFrameworkCore.MySql;
-using Microsoft.AspNetCore.DataProtection;
-using System.IO;
-using EquilibriumCore.Areas.Identity.Data;
-using EquilibriumCore.Service;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using System.Net;
+using Microsoft.Extensions.Hosting;
+using System;
+using System.Runtime.InteropServices;
+using System.Threading.Tasks;
 
 namespace EquilibriumCore
 {
@@ -46,70 +37,61 @@ namespace EquilibriumCore
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
 
-            services.Configure<IdentityOptions>(options =>
-            {
-                options.Password.RequireDigit = true;
-                options.Password.RequiredLength = 5;
-                options.Password.RequireLowercase = false;
-                options.Password.RequireNonAlphanumeric = false;
-                options.Password.RequireUppercase = false;
-            });
-
-            var environment = services.BuildServiceProvider().GetRequiredService<IHostingEnvironment>();
+            services.AddControllersWithViews();
 
 
-            services.AddDataProtection()
-                    .SetApplicationName($"my-app-{environment.EnvironmentName}")
-                    .PersistKeysToFileSystem(new DirectoryInfo($@"{environment.ContentRootPath}\keys"));
-
-           
-            services.AddDbContext<DataContext>(option=> option.UseMySql(Configuration.GetConnectionString("DefaultConnection")));
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddDbContext<DataContext>(option => option.UseMySql(Configuration.GetConnectionString("DefaultConnection")));
+            //services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
             services.AddJsReport(new LocalReporting()
-        .UseBinary(RuntimeInformation.IsOSPlatform(OSPlatform.Windows)  ?
-                    jsreport.Binary.JsReportBinary.GetBinary() :
-                    jsreport.Binary.Linux.JsReportBinary.GetBinary())
-        .AsUtility()
-        .Create());
+                .UseBinary(RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ?
+                            jsreport.Binary.JsReportBinary.GetBinary() :
+                            jsreport.Binary.Linux.JsReportBinary.GetBinary())
+                .AsUtility()
+                .Create());
             services.AddResponseCaching();
             services.AddScoped<IToolTipService, ToolTipService>();
-
+            services.AddRazorPages();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env ,IServiceProvider serviceProvider)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IServiceProvider serviceProvider)
         {
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
                 app.UseHttpsRedirection();
-                
+
             }
             else
             {
                 app.UseDeveloperExceptionPage();
-             
+
             }
+
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseAuthentication();           
+
+            app.UseRouting();
+            app.UseAuthentication();
+            app.UseAuthorization();
             app.UseCookiePolicy();
-          
+
             app.UseResponseCaching();
-            app.UseMvc(routes =>
+            app.UseEndpoints(routes =>
             {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                routes.MapControllerRoute("default", "{controller=Home}/{action=Index}/{id?}");
+                routes.MapRazorPages();
+
             });
 
             using (var serviceScope = app.ApplicationServices.GetRequiredService<IServiceScopeFactory>().CreateScope())
-	    {
-		var context = serviceScope.ServiceProvider.GetService<DataContext>();
-		context.Database.Migrate();
-                var securcontext =    serviceScope.ServiceProvider.GetService<SecurityContext>();
+            {
+                var context = serviceScope.ServiceProvider.GetService<DataContext>();
+                context.Database.Migrate();
+                var securcontext = serviceScope.ServiceProvider.GetService<SecurityContext>();
                 securcontext.Database.Migrate();
-	    }
+            }
             createAdmin(serviceProvider).GetAwaiter().GetResult();
         }
         private async Task createAdmin(IServiceProvider serviceProvider)
@@ -121,7 +103,7 @@ namespace EquilibriumCore
             if (!roleCheck)
             {
                 var roleResult = await roleManager.CreateAsync(new IdentityRole("Admin"));
-               
+
             }
 
             var adminUser = await userManager.FindByNameAsync("Admin");
