@@ -35,7 +35,7 @@
                     <td>
 
 
-                        </span>
+                        
                         <dl class="dl-horizontal" align="left">
                             <dt>
                                 @Html.DisplayNameFor(model => model.Race)
@@ -420,7 +420,7 @@
         <div style="page-break-before:always">
             <hr />
             <div id="Edition" style="display:  none ">
-                <a asp-action="Select" asp-controller="Skills" asp-all-route-data="@{ new Dictionary<string, string> { { "FP", Model.ID.ToString() } }; } ">Add a new passive</a>
+                <a asp-action="Select" asp-controller="Skills" asp-all-route-data=' new Dictionary<string, string> { { "FP", Model.ID.ToString() } }  '>Add a new passive</a>
             </div>
             <h3>Passive</h3><br />
             <input type="text" class="hidden" id="skills" name="skills" value="@Model.skills" />
@@ -430,8 +430,8 @@
 
                     <h4> @s.Name </h4> @s.CurrentLevel @:/ @s.levelMax <br />
                     <div id="Edition" style="display:  none ">
-                        <a class="btn btn-danger" asp-action="RemoveSkillsFP" asp-all-route-data="@{ new Dictionary<string, string> { {"SkillId" ,s.ID.ToString() }, {"id", Model.ID.ToString() } }; } ">-</a>
-                        <a asp-action="AddSkillsFP" asp-all-route-data="@{ new Dictionary<string, string> { {"SkillId" ,s.ID.ToString() }, {"id", Model.ID.ToString() } }; } " class="btn btn-success">+</a>
+                        <a class="btn btn-danger" asp-action="RemoveSkillsFP"  asp-all-route-data=' new Dictionary<string, string> { {"SkillId" ,s.ID.ToString() }, {"id", Model.ID.ToString() } } '>-</a>
+                        <a asp-action="AddSkillsFP" asp-all-route-data=' new Dictionary<string, string> { {"SkillId" ,s.ID.ToString() }, {"id", Model.ID.ToString() } };   ' class="btn btn-success">+</a>
                     </div>
                     @Html.Raw(ToolTips.InsertTooltip(s.currentEffect))
 
